pyproject.toml
[project]
name = "attendance-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.2.1",
    "cmake>=3.31.6",
    "cryptography>=44.0.1",
    "django>=5.1.6",
    "django-dump-load-utf8>=0.0.4",
    "djangorestframework>=3.15.2",
    "djlint>=1.36.4",
    "dlib>=19.24.6",
    "face-recognition>=1.3.0",
    "icecream>=2.1.4",
    "mypy>=1.15.0",
    "opencv-contrib-python>=4.11.0.86",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "pylint>=3.3.4",
    "pylint-django>=2.6.1",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "ruff>=0.9.6",
    "selenium>=4.29.0",
    "setuptools>=75.8.2",
]
